[refactor] next page 파일 관리 방법 변경 - 기존 방식으로 변경
