+++ src/pages/index.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 
-import Home from './index.page';
+import Home from '.';
 
 describe('Home', () => {
   it('renders a headig', () => {
